+++ timer/timer.go
@@ -11,11 +11,11 @@ const (
 var (
 	fps                 int
 	frames              int
-	current_time        float64
-	previous_time       float64
-	previous_fps_update float64
-	dt                  float64
-	average_delta       float64
+	current_time        float32
+	previous_time       float32
+	previous_fps_update float32
+	dt                  float32
+	average_delta       float32
 )
 
 func Step() {
@@ -26,18 +26,18 @@ func Step() {
 
 	time_since_last := current_time - previous_fps_update
 	if time_since_last > fps_update_frequency {
-		fps = int((float64(frames) / time_since_last) + 0.5)
-		average_delta = time_since_last / float64(frames)
+		fps = int((float32(frames) / time_since_last) + 0.5)
+		average_delta = time_since_last / float32(frames)
 		previous_fps_update = current_time
 		frames = 0
 	}
 }
 
-func GetTime() float64 {
-	return float64(sdl.GetTicks()) / 1000.0
+func GetTime() float32 {
+	return float32(sdl.GetTicks()) / 1000.0
 }
 
-func GetDelta() float64 {
+func GetDelta() float32 {
 	return dt
 }
 
@@ -45,6 +45,6 @@ func GetFPS() int {
 	return fps
 }
 
-func GetAverageDelta() float64 {
+func GetAverageDelta() float32 {
 	return average_delta
 }
